pyproject.toml
[project]
name = "octa-graph-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "h5py>=3.13.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "matplotlib>=3.10.3",
    "natsort>=8.4.0",
    "nibabel>=5.3.2",
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pip>=25.1.1",
    "scikit-image>=0.25.2",
    "scipy>=1.15.2",
    "tqdm>=4.67.1",
]
